// src/Table/lib/QuickFilter/reducer/reducer.d.ts
// eslint-disable-next-line no-unused-vars
import { ColumnProps } from '../../../../typings'

interface FilterProps {
  filterIndex: number
  property: null | string
  type: string | null
  value: any
}
type QuickFilterProps = FilterProps & ColumnProps
interface QuickFilterState {
  filters: Array<QuickFilterProps>
}
type QuickFilterAction =
  | {
      type: 'ADD_FILTER'
      payload: {
        property: null | string
        type: string | null
        value: null | string | boolean | number | Date | undefined
      } & ColumnProps
    } // typescript union types allow for leading |'s to have nicer layout
  | { type: 'REMOVE_FILTER'; payload: Array<QuickFilterProps> }
  | { type: 'UPDATE_FILTER'; payload: QuickFilterProps }
  | { type: 'ADD_OR_UPDATE_SEARCH'; payload: any }
  | { type: 'RESET' }
  | { type: 'REINITIALIZE_FILTER'; payload: Array<QuickFilterProps> }

export { QuickFilterAction, QuickFilterProps, QuickFilterState, FilterProps }